requirements.txt
MetaTrader5>=5.0.45
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
streamlit>=1.38.0
resend>=2.0.0
